setup test php
