--- api/config/packages/dev/hautelook_alice.yaml
@@ -1,2 +1,0 @@
-hautelook_alice:
-    fixtures_path: fixtures
