--- api/config/packages/test/hautelook_alice.yaml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: ../dev/hautelook_alice.yaml }
