--- api/tests/Api/BookTest.php
@@ -1,64 +1,0 @@
-<?php namespace App\Tests;
-
-use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\ApiTestCase;
-use App\Entity\Book;
-use Hautelook\AliceBundle\PhpUnit\RefreshDatabaseTrait;
-
-class BooksTest extends ApiTestCase
-{
-    // This trait provided by HautelookAliceBundle will take care of refreshing the database content to a known state before each test
-    use RefreshDatabaseTrait;
-
-    public function testGetCollection(): void
-    {
-        // The client implements Symfony HttpClient's `HttpClientInterface`, and the response `ResponseInterface`
-        $response = static::createClient()->request('GET', '/books');
-
-        $this->assertResponseIsSuccessful();
-        // Asserts that the returned content type is JSON-LD (the default)
-        $this->assertResponseHeaderSame('content-type', 'application/ld+json; charset=utf-8');
-
-        // Asserts that the returned JSON is a superset of this one
-        $this->assertJsonContains([
-            '@context' => '/contexts/Book',
-            '@id' => '/books',
-            '@type' => 'hydra:Collection',
-            'hydra:totalItems' => 100,
-            'hydra:view' => [
-                '@id' => '/books?page=1',
-                '@type' => 'hydra:PartialCollectionView',
-                'hydra:first' => '/books?page=1',
-                'hydra:last' => '/books?page=4',
-                'hydra:next' => '/books?page=2',
-            ],
-        ]);
-
-        // Because test fixtures are automatically loaded between each test, you can assert on them
-        $this->assertCount(30, $response->toArray()['hydra:member']);
-
-        // Asserts that the returned JSON is validated by the JSON Schema generated for this resource by API Platform
-        // This generated JSON Schema is also used in the OpenAPI spec!
-        $this->assertMatchesResourceCollectionJsonSchema(Book::class);
-    }
-
-    public function testCreateBook(): void
-    {
-        $response = static::createClient()->request('POST', '/books', ['json' => [
-            'name' => 'The Handmaid\'s Tale',
-            'author' => 'Margaret Atwood',
-            'datePublished' => '1985-07-31T00:00:00+00:00',
-        ]]);
-
-        $this->assertResponseStatusCodeSame(201);
-        $this->assertResponseHeaderSame('content-type', 'application/ld+json; charset=utf-8');
-        $this->assertJsonContains([
-            '@context' => '/contexts/Book',
-            '@type' => 'Book',
-            'name' => 'The Handmaid\'s Tale',
-            'author' => 'Margaret Atwood',
-            'datePublished' => '1985-07-31T00:00:00+00:00',
-        ]);
-        $this->assertMatchesRegularExpression('~^/books/\d+$~', $response->toArray()['@id']);
-        $this->assertMatchesResourceItemJsonSchema(Book::class);
-    }
-}
