+++ api/tests/Api/BooksTest.php
@@ -1,0 +1,61 @@
+<?php namespace App\Tests;
+
+use ApiPlatform\Symfony\Bundle\Test\ApiTestCase;
+use App\Entity\Book;
+use Hautelook\AliceBundle\PhpUnit\RefreshDatabaseTrait;
+
+class BooksTest extends ApiTestCase
+{
+    public function testGetCollection(): void
+    {
+        // The client implements Symfony HttpClient's `HttpClientInterface`, and the response `ResponseInterface`
+        $response = static::createClient()->request('GET', '/books');
+        $this->assertResponseIsSuccessful();
+        // Asserts that the returned content type is JSON-LD (the default)
+        $this->assertResponseHeaderSame('content-type', 'application/ld+json; charset=utf-8');
+        // // Asserts that the returned JSON is a superset of this one
+        $this->assertJsonContains([
+            '@context' => '/contexts/Book',
+            '@id' => '/books',
+            '@type' => 'hydra:Collection',
+            'hydra:member' => [
+                [
+                    "@id" => "/books/1",
+                    "@type" => "Book",
+                    "id" => 1,
+                    "rating" => 0,
+                    "body" => "string",
+                    "author" => "string",
+                    "publication_date" => "2022-08-03T00:00:00+00:00",
+                    "publicationDate" => "2022-08-03T00:00:00+00:00"
+                ]
+            ],
+        ]);
+
+        // Because test fixtures are automatically loaded between each test, you can assert on them
+        // $this->assertCount(30, $response->toArray()['hydra:member']);
+
+        // Asserts that the returned JSON is validated by the JSON Schema generated for this resource by API Platform
+        // This generated JSON Schema is also used in the OpenAPI spec!
+        $this->assertMatchesResourceCollectionJsonSchema(Book::class);
+    }
+
+    public function testCreateBook(): void
+    {
+        static::createClient()->request('POST', '/books', ['json' => [
+            "rating" => 0,
+            "body" => "string",
+            "author" => "string",
+            "publicationDate" => "2022-08-03"
+        ]]);
+
+        $this->assertResponseStatusCodeSame(201);
+        $this->assertJsonContains([
+            "@context"=> "/contexts/Book",
+            "@type"=> "Book",
+            "rating"=> 0,
+            "body"=> "string",
+            "author"=> "string",
+        ]);
+    }
+}
